requirements.txt
appdirs==1.4.4
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
cssselect==1.1.0
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
django-rest-authtoken==2.1.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
fake-useragent==0.1.11
feedparser==6.0.8
finance-datareader==0.9.31
idna==3.2
importlib-metadata==4.8.1
lxml==4.6.3
numpy==1.21.2
pandas==1.3.3
parse==1.19.0
psycopg2==2.9.1
pyee==8.2.2
PyJWT==1.7.1
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-file==1.5.1
requests-html==0.10.0
selenium==3.141.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.7
w3lib==1.22.0
websockets==9.1
yahoo-fin==0.8.9.1
zipp==3.6.0
